Move iNoNorm from Generic to Snitch since it's only used there
