+++ Deeploy/Targets/Snitch/Templates/iNoNormTemplate.py
@@ -4,14 +4,7 @@
 
 from Deeploy.DeeployTypes import NodeTemplate
 
-
-class _iNoNormTemplate(NodeTemplate):
-
-    def __init__(self, templateStr):
-        super().__init__(templateStr)
-
-
-referenceTemplate = _iNoNormTemplate("""
+referenceTemplate = NodeTemplate("""
 // iNoNorm (Name: ${nodeName}, Op: ${nodeOp})
 SnitchiNoNorm_s${data_in_type.referencedType.typeWidth}_s${data_out_type.referencedType.typeWidth}(${data_in}, ${data_out}, ${weights}, ${bias}, ${size}, ${mul}, ${log2D});
 """)
